Inline dead-cart cleanup into zombie killer; dedup test helper
- deleteDeadCartsService was single-use indirection (only the zombie killer
  called it); inline the detection + cascade-delete directly in the worker,
  which already orchestrates cross-collection cleanup. deleteCartService stays
  (genuinely reused by deleteUser and removeOrder).
- Extract the duplicated work-poll-to-completion logic from both GC tests into
  a shared runWorkToCompletion helper in tests/helpers.js.
