+++ packages/core/src/services/index.ts
@@ -16,7 +16,6 @@ import { updateCalculationService } from './updateCalculation.ts';
 import { supportedDeliveryProvidersService } from './supportedDeliveryProviders.ts';
 import { deleteUserService } from './deleteUser.ts';
 import { deleteCartService } from './deleteCart.ts';
-import { deleteDeadCartsService } from './deleteDeadCarts.ts';
 import { supportedPaymentProvidersService } from './supportedPaymentProviders.ts';
 import { supportedWarehousingProvidersService } from './supportedWarehousingProviders.ts';
 import { createEnrollmentFromCheckoutService } from './createEnrollmentFromCheckout.ts';
@@ -106,7 +105,6 @@ export default function initServices(modules: Modules, customServices: CustomSer
       >,
       migrateOrderCarts: migrateOrderCartsService as Bound<typeof migrateOrderCartsService>,
       deleteCart: deleteCartService as Bound<typeof deleteCartService>,
-      deleteDeadCarts: deleteDeadCartsService as Bound<typeof deleteDeadCartsService>,
       nextUserCart: nextUserCartService as Bound<typeof nextUserCartService>,
       findOrInitCart: findOrInitCartService as Bound<typeof findOrInitCartService>,
       initCartProviders: initCartProvidersService as Bound<typeof initCartProvidersService>,
